Extracted theme watcher to module
